release v1.2.0

+++ modules/img_transform/add_weighted/src/add_weighted.cpp
@@ -51,7 +51,7 @@ int add_weighted(
         src1.type() != FCVImageType::PLA_RGB_U8 &&
         src1.type() != FCVImageType::PKG_BGR_U8 &&
         src1.type() != FCVImageType::PKG_RGB_U8)) {
-        LOG_ERR("The input type is not surpport, which is %d \n", src1.type());
+        LOG_ERR("The input type is not surpport, which is %d \n", (int)src1.type());
         return -1;
     }
